% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Simulated Bank Data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
  \item{arrival_time}{Numeric vector of the time each customer arrives, in seconds.}
  \item{service_time}{Numeric vector of the service time for each customer, in seconds.}
}
}
\usage{
bank
}
\description{
This dataset contains simulated data on 100 customers arriving at a bank
and their corresponding service times.
}
\examples{
data(bank)
head(bank)
}
\keyword{datasets}
