% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver Function}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A numeric vector of arrival times.}

\item{ServiceTimes}{A numeric vector of service times.}

\item{NumServers}{An integer for the number of servers (default is 1).}
}
\value{
A tibble containing arrival times, service start times, chosen server, and service end times.
}
\description{
A function to simulate a multiserver queue system.
}
\details{
Multiserver
}
